js 33

+++ src/main/java/com/study/lastlayer/weekhistory/WeekHistoryDto.java
@@ -4,15 +4,15 @@ import java.time.LocalDate;
 
 public class WeekHistoryDto {
 	
-	public static record WeekHistoryRequest(String date, Double weight) {}
-	public static record WeekHistoryUpdateRequest(Double weight) {}
-	public static record WeekHistoryResponse(Long id, String date, Double weight, String message) {
+	public static record WeekHistoryRequest(String date, Float weight) {}
+	public static record WeekHistoryUpdateRequest(Float weight) {}
+	public static record WeekHistoryResponse(Long id, String date, Float weight, String message) {
 		public WeekHistoryResponse(WeekHistory entity, String message) {
 			this(entity.getId(), entity.getDate().toString(), entity.getWeight(), message);
 		}
 	}
 	
-	public static record WeekHistoryListResponse(Long id, String date, Double weight) {
+	public static record WeekHistoryListResponse(Long id, String date, Float weight) {
 		public WeekHistoryListResponse(WeekHistory entity) {
 			this(entity.getId(), entity.getDate().toString(), entity.getWeight());
 		}
